requirements.txt
Flask==1.0.2
flask-script==2.0.6
blueprint==3.4.2
flask-sqlalchemy==2.3.2
flask-mongoengine==0.9.5